Move get_saved_sample_data_for_theme() to helpers_for_testing.py
