+++ .github/scripts/tests/test_themes.py
@@ -1,6 +1,4 @@
 import os
-from pathlib import Path
-from typing import Tuple
 
 import approvaltests
 from approvaltests import verify_as_json, verify, Options
@@ -8,10 +6,9 @@ from approvaltests.scrubbers import create_regex_scrubber, Scrubber
 from approvaltests.storyboard import StoryBoard
 
 import utils
-from tests.helpers_for_testing import verify_as_markdown
+from tests.helpers_for_testing import verify_as_markdown, get_saved_sample_data_for_theme
 from tests.test_templates import JINJA_TEMPLATES_DIR
-from themes import get_theme_downloads, collect_data_for_theme, ThemeList, collect_data_for_theme_and_css, Theme, \
-    ThemeDownloads
+from themes import get_theme_downloads, collect_data_for_theme, ThemeList, collect_data_for_theme_and_css
 from utils import THEMES_JSON_FILE, get_json_from_github
 
 
@@ -95,20 +92,6 @@ def test_rendering_of_theme() -> None:
     verify_as_markdown(
         file_content,
         options=Options().with_scrubber(make_download_badge_numbers_stable()))
-
-
-def get_saved_sample_data_for_theme(theme_name: str) -> Tuple[Theme, str, ThemeDownloads]:
-    sample_data_for_theme = Path(__file__).parent.absolute() / 'sample_data/themes' / theme_name
-    assert sample_data_for_theme.exists()
-
-    theme_downloads = utils.get_json_from_file(str(sample_data_for_theme / 'stats-theme.json'))
-    theme_list: ThemeList = utils.get_json_from_file(str(sample_data_for_theme / 'community-css-themes.json'))
-    theme = theme_list[0]
-
-    with open(sample_data_for_theme / 'obsidian.css') as f:
-        css_file = f.read()
-
-    return theme, css_file, theme_downloads
 
 
 def verify_theme_data(theme_name: str) -> None:
